Simplify `Debug` display of `DebugName` (#21870)
# Objective

As noted in #21856, the `Debug` output of `DebugName` is too verbose. It
is supposed to be a thin wrapper around a string, but it renders as a
`struct`.

## Solution

Manually `impl Debug for DebugName` and write the string directly.  

## Showcase

The following code

```rust
#[derive(Debug)]
struct TestStruct {
    debug_name: DebugName,
}
let test_struct = TestStruct {
    debug_name: DebugName::type_name::<TestStruct>(),
};
println!("{test_struct:#?}");
```

Prints the following before this change

```
TestStruct {
    debug_name: DebugName {
        name: "crate_name::TestStruct",
    },
}
```

And the following after it

```
TestStruct {
    debug_name: "crate_name::TestStruct",
}
```

When the `debug` feature is disabled, it prints the following both
before and after the change

```
TestStruct {
    debug_name: DebugName,
}
```

---------

Co-authored-by: François Mockers <francois.mockers@vleue.com>