Fix extended_material_bindless example wgpu validation error on MacOS/Metal (#21729)
# Objective

- Fixes https://github.com/bevyengine/bevy/issues/21719

## Solution

- Implementing `bindless_supported` of the `AsBindGroup` trait for
`ExtendedMaterial`:
```rust
    fn bindless_supported(render_device: &RenderDevice) -> bool {
        B::bindless_supported(render_device) && E::bindless_supported(render_device)
    }
```
Before it just returned the default (true), leading to the mismatch.

I made small, mostly stylistic, improvements to `fn
bind_group_layout_entries` and a few other locations like a comment in
`extended_material` during debugging but I could move that out into a
separate PR as well. Since the changes are rather limited and related to
the issue I decided to leave it in for now.

## Testing

- Tested on Windows/NixOS and MacOS, tested multiple examples but most
testing happened on the `extended_material_bindless` and
`extended_material` examples.