Implement the infrastructure needed to support portals and mirrors.
Bevy currently supports multiple cameras and rendering to off-screen
render targets, so one might naïvely think that the engine has support
for portals and mirrors already. However, Bevy is missing two key
features that enable portals and mirrors at present:

1. Bevy has support for neither custom clip planes nor oblique clip
   planes. This prevents the construction of proper portals or mirrors,
   as meshes that intersect the portal plane must be clipped to render
   properly.

2. Bevy has no support for cameras that invert the culling mode, so
   meshes that are reflected across a plane will render inside-out.

This PR addresses the two issues above:

1. This commit introduces a new feature on `PerspectiveProjection`,
   `near_normal`. This allows the application to specify the normal of
   the near plane.  Together with adjusting the near plane distance,
   this fully enables [Lengyel oblique clipping], which is the most
   optimal way to achieve a custom near clipping plane. It allows us to
   avoid having to support custom clip planes, which are often
   implemented inefficiently in hardware.

2. This patch adds a new field on the `Camera` component,
   `invert_culling`. This field causes the Bevy renderer to invert the
   front face setting when rendering the objects visible from that
   camera. When coupled with an appropriately-set [Householder matrix]
   on the camera, this allows correct rendering of objects reflected
   across a plane.

Additionally, this PR adds a new function to `bevy_math::mat3`,
`reflection_matrix`. This generates the matrix that reflects objects
across a plane, suitable for encoding into a `Transform`. It's fully
documented for ease of use.

Finally, a new example, `mirror`, has been added. This example is a
complete instance of a working mirror, combining a camera with a
Householder matrix, oblique projection, and inverted culling with a
custom material to render an animated mesh and its planar reflection.
The camera and mesh may be moved with the mouse, and the off-screen
render target that stores the rendered contents of the mirror world is
properly resized when the user resizes the window.

[Lengyel oblique clipping]: https://terathon.com/lengyel/Lengyel-Oblique.pdf

[Householder matrix]: https://en.wikipedia.org/wiki/Householder_transformation