Release the asset infos lock before acquiring the file transaction lock. (#22017)
# Objective

- Fix the first flake related to #22001.
- We had a double lock problem. We would first lock `asset_infos`, then
lock one of the assets within `asset_infos` **and then dropped the
`asset_infos` lock**. This means however that if a process needs to
access `asset_infos` (which is necessary for loading nested assets
during processing), we'll have one thread trying to lock 1) asset_infos,
2) per-asset lock, and we'll have another thread trying to lock 1)
per-asset lock (we've already dropped the asset_infos lock), 2)
asset_infos. A classic deadlock!

## Solution

- Before locking the per-asset lock, we clone the `Arc<RwLock>` out of
the `asset_infos`, drop the `asset_infos`, and only then lock the
per-asset lock. This ensures that we never hang on the `asset_infos`
lock when just trying to fetch a single asset.
- Make all the access to `asset_infos` "short" - they get what they need
out of `asset_infos` and then drop the lock as soon as possible.
- I also used `?` to cleanup some methods in `ProcessorAssetInfos`,
where previously it was "remove the asset info, then have one big `if
let` after it". Now we just return early if the value is none.
- I also happened to fix a weird case where the new path in a rename
wasn't guarded by a transaction lock. Now it is!

## Testing

- Running without this fix I get "Ran out of loops" from the
`only_reprocesses_wrong_hash_on_startup` test quite quickly (after a
minute or so). With this fix, I now only get the assertion failure
problem. If I also skip that assertion, the test hasn't flaked for a
while! Yay, no more deadlock!