Increase the number of supported clustered decals if the `PARTIALLY_BOUND_BINDING_ARRAY` `wgpu` feature is present. (#22028)
The current clustered decal code predates
`PARTIALLY_BOUND_BINDING_ARRAY` support in `wgpu`. As a result, only 8
clustered decals are allowed in the scene in order to avoid paying a
high cost even if no decals of that type are present. This limitation
makes Bevy's current clustered decal support insufficient for many use
cases (e.g. bullet holes), so this patch increases the limit from 8 to
1024 if the `wgpu` features include `PARTIALLY_BOUND_BINDING_ARRAY`.

---------

Co-authored-by: François Mockers <francois.mockers@vleue.com>