Change a function system to be exclusive (#21742)
# Objective

- I was experimenting with a library for combining multiple systems into
one system that could be scheduled, but there wasn't a way of marking a
function system run exclusively.

## Solution

- Add a method to `SystemMeta` to set the exclusive flag. Note that this
PR doesn't add anyway to actually set the flag easily. You would have to
create a custom system param to be able to set it.
- Add a system param to mark a function system to run exclusively.

## Testing

- Haven't tested. The one worry would be how it interacts with the
executor, but in theory this should just work, since the flag is the
only way the executor knows that a system is exclusive or not.

---------

Co-authored-by: Alice Cecile <alice.i.cecile@gmail.com>
Co-authored-by: Christian Hughes <9044780+ItsDoot@users.noreply.github.com>