Fix immediate nested loaded assets not loading their dependencies. (#21863)
# Objective

- Previously if you loaded an immediate nested asset, it would **not**
load any of its dependencies. This means that you could have handles in
your asset that would never load.
- Fixes #20988

## Solution

- Propagate loading dependencies to immediate asset loads.
- I also deduped some test App setup code.

## Testing

- Added a test.

One case that does have a slightly undesirable effect is when you
immediately-load an asset, take some data from it (but no dependency
handles), and then drop the immediately-loaded asset. This change will
now also initiate a dependency load, even though we won't use it.

However, A) it is much more surprising to have a handle that will never
be loaded, B) we should end up cancelling the dependency load once we
drop the handle, so we shouldn't pay too much of the cost of loading.