Feathers: add override cursor (#21824)
# Objective

Fixes #21801 and tweaks the feathers example to exhibit the new
behavior.

## Solution

What the issue suggested. Added a new resource and adjusted the
`EntityCursor` interaction logic in `cursor.rs`.

## Testing

Tested it via the example tweak where the wide button now toggles the
override logic.
(Note: I wasn't able to get the load cursor to show up on macOS, but I
think this is a `winit` issue: `cursor `busyButClickableCursor` appears
to be invalid`)

---------

Co-authored-by: Alice Cecile <alice.i.cecile@gmail.com>