increase morph targets to 256 (#21421)
# Objective

our current morph target / blend shape limit is 64 per model, documented
as "to support all hardware". many assets use more than this ("over a
hundred" is common), so i'd like to increase this limit.

the relevant hardware constraint is uniform buffer size (usage in
morph.wgsl). morph targets use 4 bytes per entry, and wgpu's downlevel
uniform buffer size limit is 16k so we have a lot of room from the 256
bytes we are currently limiting to, while still supporting the vast
majority of hardware.

## Solution

increase the limit to 256 targets / 1k.

we also increase the input to the buffer to this 1k size: this was not
done previously but worked as long as the min buffer alignment was 256
(which it probably always is at the moment) as it was sized up to match
alignment previously.

there may be a slight performance impact from using a larger size in all
cases but it should be negligible.