UI per text section picking (#22047)
# Objective

Allow picking text per section.

Fixes #17477

## Solution

In `ui_picking` whenever a node with text components is below a pointer,
don't add the `Text` node's entity to the `hit_nodes` list. Instead,
iterate the list of text section bounding rects in its `TextLayoutInfo`
component to find which text section the pointer is over, and add that
section's entity id to the `hit_nodes` list.

* Add optional text components to `NodeQuery`.
* New helper function, `pick_ui_text`. Finds if a text section is
hovered and returns its id.
* Add the entities of text sections to `hit_nodes` when hovered.
* Add the target camera entity to `hit_nodes`, instead of querying for
it a second time.
* When a picked node is a text node and `require_markers` is enabled,
check if the hit section entity has a `Pickable` component, instead of
the text node.

#

This design might need to be changed later to allow for text selection
and copying. But there doesn't seem any need to overcomplicate things
for now by considering it.

## Testing

I added observers to the `TextBackgroundColors` example's text entities.
The observers turn a text section's text color to white when the pointer
hovers them.

```
cargo run --example text_background_colors
```